[SPARK-47337][SQL][DOCKER] Upgrade DB2 Docker version to 11.5.8.0
